+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - Live Preview display with Channel Entries Tag replacement
 - Error where GraphQL Tag queries accumulated arguments
 - Fluid field database connection issue when default Laravel db connection is not configured
+- Better handling of FieldContent methods when value is null
 
 ### Changed